+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 * Bump stripes-components dependency to v2.0.5. Makes filters work in nested apps. Fixes UIU-430.
 * Pass `props.browseOnly` through to `<SearchAndSort>` to optionally disable user creation and editing. Refs UIPFU-6. Available from v2.12.2.
 * Get alternate fixed renewal period from loan policy. Fixes UIU-433.
+* Get alternate rolling renewal period from loan policy. Fixes UIU-434 and UIU-435.
 * Remove <Autocomplete> - Use new default `country` field control for address fields. Fixes UIU-298.
 
 ## [2.12.0](https://github.com/folio-org/ui-users/tree/v2.12.0) (2017-11-28)